hide logging options in RouteTabs
