+++ src/components/routes/RoutesApps.vue
@@ -31,7 +31,6 @@ import { i18n } from '../../boot/i18n';
 import { rideToWorkByBikeConfig } from '../../boot/global_vars';
 
 // stores
-import { useRegisterChallengeStore } from '../../stores/registerChallenge';
 import { useStravaStore } from 'src/stores/strava';
 import { useTripsStore } from '../../stores/trips';
 
@@ -51,7 +50,6 @@ export default defineComponent({
     const urlGooglePlay = rideToWorkByBikeConfig.urlGooglePlay;
     const route = useRoute();
     const router = useRouter();
-    const registerChallengeStore = useRegisterChallengeStore();
     const stravaStore = useStravaStore();
     const tripsStore = useTripsStore();
     const isLoadingAuthStrava = ref<boolean>(false);
@@ -113,10 +111,6 @@ export default defineComponent({
       return result;
     });
 
-    const isCurrentUserApproved = computed<boolean>((): boolean => {
-      return registerChallengeStore.getIsCurrentUserApproved;
-    });
-
     return {
       apps,
       enabledAppsForManualLogging,
@@ -124,7 +118,6 @@ export default defineComponent({
       isLoadingAuthStrava,
       urlAppStore,
       urlGooglePlay,
-      isCurrentUserApproved,
     };
   },
 });
@@ -138,17 +131,11 @@ export default defineComponent({
       <section-heading class="q-mb-md" data-cy="routes-apps-title-auto">
         {{ $t('routes.titleAutomaticLogging') }}
         <template #perex>
-          <div
-            v-if="isCurrentUserApproved"
-            v-html="$t('routes.hintAutomaticLogging')"
-          />
-          <q-banner v-else class="bg-warning text-gray-10 rounded-borders">
-            {{ $t('routes.hintLoggingNotApproved') }}
-          </q-banner>
+          <div v-html="$t('routes.hintAutomaticLogging')" />
         </template>
       </section-heading>
       <!-- App banners -->
-      <div v-if="isCurrentUserApproved" class="flex column gap-16">
+      <div class="flex column gap-16">
         <banner-routes-app
           v-for="app in apps"
           :app="app"
