+++ src/i18n/cs.toml
@@ -1008,7 +1008,8 @@ buttonSaveChangesCount = "Žádné změny | Uložit {count} změnu | Uložit {co
 buttonTraceMap = "Zakreslit do mapy"
 descriptionRoutes = "<p>Jízdy můžete zapisovat až 8 dní zpětně.</p><p>Pokud kombinujete dopravní prostředky, zapište tu nejdelší část, případně druhou nejdelší je-li na vlastní pohon.</p>"
 hintAutomaticLogging = "Pro automatický záznam můžete využít jednu z těchto aplikací třetích stran.<br />(Ověřit, zda jste propojeni s Cyclers či Na kole Prahou můžete z technických důvodů pouze přímo v aplikacích.)"
-hintLoggingNotApproved = "Nejste schválen*a jako člen*ka týmu. Pro propojení svého účtu s aplikacemi požádejte své kolegy o schválení."
+hintAutomaticLoggingNotApproved = "Nejste schválen*a jako člen*ka týmu. Pro propojení svého účtu s aplikacemi požádejte své kolegy o schválení."
+hintManualLoggingNotApproved = "Nejste schválen*a jako člen*ka týmu. Pro aktivování zápisu cest požádejte své kolegy o schválení."
 hintLockedSection = "Tato sekce s vám otevře ihned, jakmile začne výzva."
 hintManualLogging = "Mimo toto webové rozhraní můžete také využívat naše oficiální mobilní aplikace:"
 hintTabApp = "Pro automatický záznam můžete využít jednu z těchto aplikací třetích stran:"
