+++ src/i18n/sk.toml
@@ -1007,7 +1007,8 @@ buttonSaveChangesCount = "Žiadne zmeny | Uložiť {count} zmenu | Uložiť {cou
 buttonTraceMap = "Zakreslit"
 descriptionRoutes = "<p>Jazdy môžete nahrávať až 8 dní vopred.</p><p>Ak kombinujete vozidlá, zaznamenajte najdlhšiu časť alebo druhú najdlhšiu časť, ak jazdíte sami.</p>"
 hintAutomaticLogging = "Na automatické nahrávanie môžete použiť jednu z týchto aplikácií tretích strán.<br />(Overiť, či ste prepojení s Cyclers či Na kole Prahou môžete z technických dôvodov len priamo v aplikáciách.)"
-hintLoggingNotApproved = "Nie ste schválený ako člen tímu. Pre prepojenie svojho účtu s aplikáciami požiadajte svojich kolegov o schválenie."
+hintAutomaticLoggingNotApproved = "Nie ste schválený ako člen tímu. Pre prepojenie svojho účtu s aplikáciami požiadajte svojich kolegov o schválenie."
+hintManualLoggingNotApproved = "Nie ste schválený ako člen tímu. Pre aktiváciu zápisu ciest požiadajte svojich kolegov o schválenie."
 hintLockedSection = "Táto časť sa otvorí hneď po začatí výzvy."
 hintManualLogging = "Okrem tohto webového rozhrania môžete používať aj naše oficiálne mobilné aplikácie:"
 hintTabApp = "Na automatické nahrávanie môžete použiť jednu z týchto aplikácií tretích strán:"
